Merge changes from neural network testing

+++ NestEngine/src/Nest/graphics/VertexArray.cpp
@@ -28,9 +28,9 @@ namespace Nest
 	{
 		bind();
 		vBuffer->bind();
-		unsigned int offset = 0;
+		size_t offset = 0;
 		const auto& layout = vbLayout->getLayout();
-		for (unsigned int i = 0; i < layout.size(); ++i)
+		for (size_t i = 0; i < layout.size(); ++i)
 		{
 			glEnableVertexAttribArray(i);
 			glVertexAttribDivisor(i, layout[i].advanceCount);
